+++ util/ptrutil/slice_header.go
@@ -10,7 +10,7 @@ type SliceHeader struct {
 	Cap  int
 }
 
-func GetSliceHeader(p unsafe.Pointer, length uint32) *SliceHeader {
+func GetSliceHeader(p unsafe.Pointer, length uint64) *SliceHeader {
 	return &SliceHeader{
 		Data: p,
 		Len:  int(length),
